Feature#27 전적 조회 기능에  소환사 정보만 추출하여 DB 저장

+++ Search-service/src/main/java/com/ggwp/searchservice/summoner/domain/Summoner.java
@@ -25,7 +25,7 @@ public class Summoner {
     @Column(name = "summoner_id")
     private String id; // summoner_id encrypted
 
-    private Long profileIconId; // 프로필 아이콘 번호
+    private int profileIconId; // 프로필 아이콘 번호
 
 //    private String accountId; // account encrypted 필요없어 보임
